requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-background-tasks==1.2.0
django-compat==1.0.15
django-filter==0.7
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.9.2
idna==2.8
Markdown==2.4.1
oauthlib==3.0.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
